moved logging to separate header
